requirements.txt
flask
python-dotenv
flask-mail
flask-wtf
mailtrap